third_party/blink/web_tests/fast/filesystem/read-directory-expected.txt

DirectoryReader.readEntries() test.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Successfully obtained Persistent FileSystem:file__0:Temporary
PASS entries.__proto__ is Array.prototype
PASS entries.__proto__ is Array.prototype
PASS readEntriesCount is entriesCallbackCount
PASS resultEntries.length is testEntriesCount
Entry:/a isDirectory:true
PASS testEntries[entry.fullPath] is entry.isDirectory
Entry:/b isDirectory:false
PASS testEntries[entry.fullPath] is entry.isDirectory
Entry:/c isDirectory:true
PASS testEntries[entry.fullPath] is entry.isDirectory
Entry:/d isDirectory:false
PASS testEntries[entry.fullPath] is entry.isDirectory
Entry:/e isDirectory:false
PASS testEntries[entry.fullPath] is entry.isDirectory
Entry:/f isDirectory:true
PASS testEntries[entry.fullPath] is entry.isDirectory
PASS successfullyParsed is true

TEST COMPLETE
